src/requirements.txt
-i https://pypi.org/simple
blinker==1.7.0; python_version >= '3.8'
cachetools==5.3.2; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
exceptiongroup==1.2.0; python_version < '3.11'
flask==3.0.0; python_version >= '3.8'
flask-httpauth==4.8.0
google-auth==2.25.2; python_version >= '3.7'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
httplib2==0.22.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.6; python_version >= '3.5'
importlib-metadata==7.0.1; python_version < '3.10'
iniconfig==2.0.0; python_version >= '3.7'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
oauth2client==4.1.3
packaging==23.2; python_version >= '3.7'
passlib==1.7.4
pep8==1.7.1
pluggy==1.3.0; python_version >= '3.8'
pyasn1==0.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyjwt==2.8.0; python_version >= '3.7'
pyparsing==3.1.1; python_version >= '3.1'
pytest==7.4.3; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.23; python_version >= '3.7'
tomli==2.0.1; python_version < '3.11'
typing-extensions==4.9.0; python_version >= '3.8'
urllib3==2.1.0; python_version >= '3.8'
werkzeug==3.0.1; python_version >= '3.8'
zipp==3.17.0; python_version >= '3.8'
